refactor(RegulatoryStatus): Change format of RegulatoryStatus Enum
